Refactor verdaccio set up into createGenesisWorkspace utility
