+++ e2e/genesis-e2e/tests/genesis.spec.ts
@@ -1,39 +1,14 @@
-import { TestingWorkspace, Verdaccio } from "@chiubaka/nx-plugin-testing";
-import { tmpProjPath, uniq } from "@nrwl/nx-plugin/testing";
-import { ensureDirSync, moveSync, removeSync } from "fs-extra";
-import os from "node:os";
+import {
+  createGenesisWorkspace,
+  TestingWorkspace,
+} from "@chiubaka/nx-plugin-testing";
 import path from "node:path";
 
 describe("genesis", () => {
-  let verdaccio: Verdaccio;
   let workspace: TestingWorkspace;
 
   beforeAll(() => {
-    verdaccio = new Verdaccio("http://localhost:4873");
-
-    const distPackagesDir = path.join(__dirname, "../../../dist/packages");
-
-    verdaccio.publish(path.join(distPackagesDir, "genesis"));
-    verdaccio.publish(path.join(distPackagesDir, "nx-plugin"));
-
-    const workspaceScope = "chiubaka";
-    const workspaceName = "genesis";
-
-    const tmpDir = path.join(os.tmpdir(), uniq(workspaceName));
-    ensureDirSync(tmpDir);
-
-    verdaccio.npx(
-      `genesis --workspace-scope=${workspaceScope} --workspace-name=${workspaceName} --registry=${verdaccio.getUrl()} --description="Test repo for genesis CLI E2E tests" --skip-github`,
-      tmpDir,
-    );
-
-    const tmpDestination = path.join(tmpDir, workspaceName);
-    const destination = path.join(tmpProjPath(), "..", workspaceName);
-
-    removeSync(destination);
-    moveSync(tmpDestination, destination);
-
-    workspace = new TestingWorkspace(destination);
+    workspace = createGenesisWorkspace("chiubaka", "genesis");
   });
 
   afterAll(async () => {
