+++ packages/nx-plugin-testing/package.json
@@ -24,7 +24,6 @@
   "homepage": "https://github.com/chiubaka/nx-plugin#readme",
   "dependencies": {
     "@nrwl/devkit": "^14.5.6",
-    "nx": "^14.5.6",
-    "portfinder": "^1.0.32"
+    "nx": "^14.5.6"
   }
 }
