+++ packages/nx-plugin-testing/src/setup/index.ts
@@ -1,1 +1,2 @@
+export * from "./createGenesisWorkspace";
 export * from "./createTestingWorkspace";
